refactor: consensus engine
