+++ bin/ress/Cargo.toml
@@ -12,9 +12,12 @@ workspace = true
 
 [dependencies]
 # ress
-ress-protocol.workspace = true
-ress-node.workspace = true
 ress-common.workspace = true
+ress-provider.workspace = true
+ress-protocol.workspace = true
+ress-network.workspace = true
+ress-engine.workspace = true
+ress-rpc.workspace = true
 ress-testing.workspace = true
 
 # alloy
@@ -25,16 +28,17 @@ alloy-rpc-types-engine.workspace = true
 alloy-provider.workspace = true
 
 # reth
-reth-rpc-api.workspace = true
+reth-chainspec.workspace = true
 reth-network.workspace = true
 reth-network-api.workspace = true
+reth-network-peers.workspace = true
 reth-node-ethereum.workspace = true
+reth-rpc-api.workspace = true
 reth-rpc-engine-api.workspace = true
-reth-chainspec.workspace = true
+reth-rpc-builder.workspace = true
 reth-consensus-debug-client.workspace = true
 reth-cli.workspace = true
 reth-ethereum-cli.workspace = true
-reth-network-peers.workspace = true
 
 # misc
 clap.workspace = true
