+++ crates/engine/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ress-node"
+name = "ress-engine"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -12,25 +12,21 @@ workspace = true
 
 [dependencies]
 # ress
-ress-common.workspace = true
-ress-network.workspace = true
-ress-rpc.workspace = true
-ress-vm.workspace = true
+ress-primitives.workspace = true
 ress-provider.workspace = true
-ress-testing.workspace = true
+ress-vm.workspace = true
+ress-network.workspace = true
 
 # reth
 reth-primitives.workspace = true
 reth-primitives-traits.workspace = true
+reth-errors.workspace = true
 reth-chainspec.workspace = true
 reth-consensus.workspace = true
 reth-node-api.workspace = true
 reth-node-ethereum.workspace = true
-reth-rpc-builder.workspace = true
-reth-trie-sparse.workspace = true
-reth-errors.workspace = true
-reth-network-peers.workspace = true
 reth-trie.workspace = true
+reth-trie-sparse.workspace = true
 reth-engine-tree.workspace = true
 reth-ethereum-engine-primitives.workspace = true
 
@@ -38,11 +34,9 @@ reth-ethereum-engine-primitives.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types-engine.workspace = true
 alloy-rlp.workspace = true
-alloy-eips.workspace = true
 
 # misc
 tokio.workspace = true
 tracing.workspace = true
-thiserror.workspace = true
 rayon.workspace = true
 itertools.workspace = true
