+++ crates/engine/src/lib.rs
@@ -1,0 +1,12 @@
+//! Ress consensus engine.
+
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+
+// TODO:
+use ress_network as _;
+
+/// Engine tree.
+pub mod tree;
+
+/// Consensus engine.
+pub mod engine;
